+++ database-connector/src/main/java/life/qbic/finance/persistence/OfferRepository.java
