+++ database-connector/src/main/java/life/qbic/finance/persistence/SimpleOfferSearchService.java
