+++ database-connector/src/main/java/life/qbic/projectmanagement/persistence/ProjectPreviewJpaRepository.java
