+++ database-connector/src/main/java/life/qbic/projectmanagement/persistence/ProjectPreviewRepository.java
