+++ database-connector/src/main/java/life/qbic/projectmanagement/persistence/ProjectRepositoryImpl.java
@@ -1,7 +1,11 @@
 package life.qbic.projectmanagement.persistence;
 
+import static life.qbic.logging.service.LoggerFactory.logger;
+
 import java.util.List;
 import java.util.Optional;
+import life.qbic.logging.api.Logger;
+import life.qbic.projectmanagement.application.ProjectCreationService;
 import life.qbic.projectmanagement.domain.project.Project;
 import life.qbic.projectmanagement.domain.project.ProjectCode;
 import life.qbic.projectmanagement.domain.project.ProjectId;
@@ -12,7 +16,7 @@ import org.springframework.stereotype.Component;
 
 
 /**
- * <b>Project JPA repository</b>
+ * <b>Project repository implementation</b>
  *
  * <p>Implementation for the {@link ProjectRepository} interface.
  *
@@ -24,25 +28,39 @@ import org.springframework.stereotype.Component;
  * dependency
  * upon creation.
  *
+ * Also handles project storage in openBIS through {@link QbicProjectDataRepo}
+ *
  * @since 1.0.0
  */
 @Component
 @Scope("singleton")
-public class ProjectJpaRepository implements ProjectRepository {
+public class ProjectRepositoryImpl implements ProjectRepository {
 
+  private static final Logger log = logger(ProjectRepositoryImpl.class);
   private final QbicProjectRepo projectRepo;
+  private final QbicProjectDataRepo projectDataRepo;
 
   @Autowired
-  public ProjectJpaRepository(QbicProjectRepo projectRepo) {
+  public ProjectRepositoryImpl(QbicProjectRepo projectRepo, QbicProjectDataRepo projectDataRepo) {
     this.projectRepo = projectRepo;
+    this.projectDataRepo = projectDataRepo;
   }
 
   @Override
   public void add(Project project) {
-    if (doesProjectExistWithId(project.getId())) {
+    ProjectCode projectCode = project.getProjectCode();
+    if (doesProjectExistWithId(project.getId()) || projectDataRepo.projectExists(projectCode)) {
       throw new ProjectExistsException();
     }
     projectRepo.save(project);
+
+    try {
+      projectDataRepo.add(project.getProjectCode());
+    } catch (Exception e) {
+      log.error("Could not add project to openBIS. Removing project from repository, as well.");
+      projectRepo.delete(project);
+      throw e;
+    }
   }
 
   @Override
