+++ database-connector/src/main/java/life/qbic/projectmanagement/persistence/repository/VocabularyCode.java
